Started refactoring user model file into new subdirectory.
